Enabled full navigation data for netcoreapp1.0
