+++ src/NUnitTestAdapterTests/NavigationDataTests.cs
@@ -52,16 +52,13 @@ namespace NUnit.VisualStudio.TestAdapter.Tests
         {
             VerifyNavigationData(suffix, methodName, "NUnitTestAdapterTests", expectedLineDebug, expectedLineRelease);
         }
-      
-#if !NETCOREAPP1_0
-        // .NET Standard does not have the assembly resolvers, so the fixes for this do not work
+
         [TestCase("+DerivedFromExternalAbstractClass", "EmptyMethod_ThreeLines", 6, 7)]
         [TestCase("+DerivedFromExternalConcreteClass", "EmptyMethod_ThreeLines", 13, 14)]
         public void VerifyNavigationData_WithExternalAssembly(string suffix, string methodName, int expectedLineDebug, int expectedLineRelease)
         {
             VerifyNavigationData(suffix, methodName, "NUnit3AdapterExternalTests", expectedLineDebug, expectedLineRelease);
         }
-#endif
 
         private void VerifyNavigationData(string suffix, string methodName, string expectedDirectory, int expectedLineDebug, int expectedLineRelease)
         {
